+++ Changelog.md
@@ -13,6 +13,11 @@
 - The `ApolloClient` instance used by a Mutation is now available in that
   Mutation's result.
   [PR #1945](https://github.com/apollographql/react-apollo/pull/1945)
+- When the `Query` `skip` prop is set to `true`, make sure the render prop
+  `loading` param is set to `false`, since we're not actually loading
+  anything.
+  [Issue #1869](https://github.com/apollographql/react-apollo/issues/1869)
+  [PR #1916](https://github.com/apollographql/react-apollo/pull/1916)
 
 ### 2.1.5